refactor(statistics): split projection repository into query services
Extract projection time-series, feature, and diagnosis query logic into dedicated services and keep AllocationStatsProjectionRepository as a delegating facade for compatibility.
